fix(test): export tests and refactor QuteProjectTest utility
Fixes #573
