+++ qute.jdt/com.redhat.qute.jdt.test/META-INF/MANIFEST.MF
@@ -25,3 +25,6 @@ Require-Bundle: org.junit,
  org.eclipse.m2e.lifecyclemapping.defaults
 Import-Package: com.google.gson
 Bundle-ActivationPolicy: lazy
+Export-Package: com.redhat.qute.jdt,
+ com.redhat.qute.jdt.java,
+ com.redhat.qute.jdt.template
