+++ qute.jdt/com.redhat.qute.jdt.test/src/main/java/com/redhat/qute/jdt/java/JavaCodeLensTest.java
@@ -11,8 +11,8 @@
 *******************************************************************************/
 package com.redhat.qute.jdt.java;
 
-import static com.redhat.qute.jdt.internal.QuteProjectTest.getJDTUtils;
-import static com.redhat.qute.jdt.internal.QuteProjectTest.loadMavenProject;
+import static com.redhat.qute.jdt.QuteProjectTest.getJDTUtils;
+import static com.redhat.qute.jdt.QuteProjectTest.loadMavenProject;
 import static org.junit.Assert.assertEquals;
 
 import java.util.Arrays;
@@ -35,7 +35,7 @@ import org.junit.Test;
 
 import com.redhat.qute.commons.QuteJavaCodeLensParams;
 import com.redhat.qute.jdt.QuteSupportForJava;
-import com.redhat.qute.jdt.internal.QuteProjectTest.QuteMavenProjectName;
+import com.redhat.qute.jdt.QuteProjectTest.QuteMavenProjectName;
 
 /**
  * Tests for Qute @CheckedTemplate support code lens inside Java files.
